requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT
shortuuid
python-dotenv